+++ version/changelog.md
@@ -8,6 +8,7 @@
 - memorize directories and enable backward history navigation
 - support nested zips
 - add context menu using SHIFT+F10
+- add open with dialog for open with native apps
 
 ### v1.5.0 (31.01.2026)